+++ src/nodepacks/makerbeam/MakerBeamNode.tsx
@@ -1,5 +1,5 @@
 import { type NodeProps } from '@xyflow/react'
-import { BaseNode, NumberInput } from '@/nodes/BaseNode'
+import { BaseNode, ExpressionInput } from '@/nodes/BaseNode'
 import { useEditorStore } from '@/store/editorStore'
 import type { MakerBeamData } from './types'
 
@@ -8,7 +8,13 @@ export function MakerBeamNode({ id, data, selected }: NodeProps) {
   const update = useEditorStore((s) => s.updateNodeData)
   return (
     <BaseNode id={id} category="makerbeam" label="makerbeam" selected={selected}>
-      <NumberInput label="length (mm)" value={d.length} min={1} step={10} onChange={(v) => update(id, { length: v })} />
+      <ExpressionInput
+        label="length (mm)"
+        value={d.length}
+        step={10}
+        min={1}
+        onChange={(v) => update(id, { length: typeof v === 'number' ? v : Number(v) || 0 })}
+      />
       <div className="text-[9px] text-yellow-600 mt-1">MakerBeamXL 15×15mm</div>
     </BaseNode>
   )
